Elaborate on CY side

+++ generators/firechip/src/main/scala/bridges/uart/UARTModule.scala
@@ -10,7 +10,7 @@ import sifive.blocks.devices.uart.UARTParams
 class UARTDUT(implicit val p: Parameters) extends Module {
   val uartParams = UARTParams(address = 0x10013000)
 
-  val ep = Module(new UARTBridge(uartParams, 1e8 / 1e6))
+  val ep = Module(new UARTBridge(uartParams, (1e8 / 1e6).toInt))
   ep.io.reset    := reset
   ep.io.clock    := clock
   ep.io.uart.txd := ep.io.uart.rxd
